Added jest setup and teardown files.
