+++ jest-setup.js
@@ -1,0 +1,3 @@
+const dotenv = require('dotenv');
+
+module.exports = async () => dotenv.config();
