+++ jest-teardown.js
@@ -1,0 +1,5 @@
+const db = require('./server/database/db');
+
+module.exports = async () => {
+  db.end();
+};
