+++ package.json
@@ -7,7 +7,7 @@
     "start": "NODE_ENV='production' node ./server/server.js",
     "build": "NODE_ENV='production' webpack",
     "dev": "NODE_ENV='development' concurrently 'kill-port 3000' 'webpack serve' 'nodemon ./server/server.js'",
-    "test": "NODE_ENV='development' jest --forceExit"
+    "test": "jest"
   },
   "repository": {
     "type": "git",
